AnnotationFunction-class {ComplexHeatmap}	R Documentation

The AnnotationFunction Class

Description

The AnnotationFunction Class

Details

The heatmap annotation is basically graphics aligned to the heatmap columns or rows. There is no restriction for the graphic types, e.g. it can be heatmap-like annotation or points. Here the AnnotationFunction class is designed for creating complex and flexible annotation graphics. As the main part of the class, it uses a user-defined function to define the graphics. It also keeps information of the size of the plotting regions of the annotation. And most importantly, it allows subsetting to the annotation to draw a subset of the graphics, which is the base for the splitting of the annotations.

See AnnotationFunction constructor for details.

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]