ColorMapping-class {ComplexHeatmap}	R Documentation

Class for Color Mapping

Description

Class for Color Mapping

Details

The ColorMapping-class handles color mapping for discrete values and continuous values. Discrete values are mapped by setting a vector of colors and continuous values are mapped by setting a color mapping function.

Methods

The ColorMapping-class provides following methods:

ColorMapping: contructor methods.

map_to_colors,ColorMapping-method: mapping values to colors.

color_mapping_legend,ColorMapping-method: draw legend or get legend as an object.

Author(s)

Zuguang Gu <z.gu@dkfz.de>

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]