[.AnnotationFunction {ComplexHeatmap}	R Documentation

Subset an AnnotationFunction Object

Description

Subset an AnnotationFunction Object

Usage

## S3 method for class 'AnnotationFunction'
x[i]

Arguments

x
An AnnotationFunction-class object.

i
A vector of indices.

Details

One good thing for designing the AnnotationFunction-class is it can be subsetted, and this is the base for the splitting of the annotations.

Examples

anno = anno_simple(1:10)
anno[1:5]
draw(anno[1:5], test = "subset of column annotation")

[Package ComplexHeatmap version 2.10.0 Index]