[.SingleAnnotation {ComplexHeatmap}	R Documentation

Subset an SingleAnnotation Object

Description

Subset an SingleAnnotation Object

Usage

## S3 method for class 'SingleAnnotation'
x[i]

Arguments

x
An SingleAnnotation-class object.

i
A vector of indices.

Details

The SingleAnnotation class object is subsetable only if the containing AnnotationFunction-class object is subsetable. All the anno_* functions are subsetable, so if the SingleAnnotation object is constructed by one of these functions, it is also subsetable.

Examples

ha = SingleAnnotation(value = 1:10)
ha[1:5]
draw(ha[1:5], test = "ha[1:5]")

[Package ComplexHeatmap version 2.10.0 Index]