Heatmap-class {ComplexHeatmap}	R Documentation

Class for a Single Heatmap

Description

Class for a Single Heatmap

Details

The Heatmap-class is not responsible for heatmap legend and annotation legends. The draw,Heatmap-method method constructs a HeatmapList-class object which only contains one single heatmap and call draw,HeatmapList-method to make the complete heatmap.

Methods

The Heatmap-class provides following methods:

Heatmap: constructor method.

draw,Heatmap-method: draw a single heatmap.

add_heatmap,Heatmap-method append heatmaps and annotations to a list of heatmaps.

row_order,HeatmapList-method: get order of rows

column_order,HeatmapList-method: get order of columns

row_dend,HeatmapList-method: get row dendrograms

column_dend,HeatmapList-method: get column dendrograms

Author(s)

Zuguang Gu <z.gu@dkfz.de>

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]