HeatmapAnnotation-class {ComplexHeatmap}	R Documentation

Class for Heatmap Annotations

Description

Class for Heatmap Annotations

Details

A complex heatmap contains a list of annotations which are represented as graphics placed on rows and columns. The HeatmapAnnotation-class contains a list of single annotations which are represented as a list of SingleAnnotation-class objects.

Methods

The HeatmapAnnotation-class provides following methods:

HeatmapAnnotation: constructor method.

draw,HeatmapAnnotation-method: draw the annotations.

Author(s)

Zuguang Gu <z.gu@dkfz.de>

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]