SingleAnnotation-class {ComplexHeatmap}	R Documentation

Class for a Single Annotation

Description

Class for a Single Annotation

Details

The SingleAnnotation-class is used for storing data for a single annotation and provides methods for drawing annotation graphics.

Methods

The SingleAnnotation-class provides following methods:

SingleAnnotation: constructor method

draw,SingleAnnotation-method: draw the single annotation.

Author(s)

Zuguang Gu <z.gu@dkfz.de>

See Also

The SingleAnnotation-class is always used internally. The public HeatmapAnnotation-class contains a list of SingleAnnotation-class objects and is used to add annotation graphics on heatmaps.

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]