add_heatmap-HeatmapList-method {ComplexHeatmap}	R Documentation

Add heatmaps and row annotations to the heatmap list

Description

Add heatmaps and row annotations to the heatmap list

Usage

## S4 method for signature 'HeatmapList'
add_heatmap(object, x, direction = c("horizontal", "vertical"))

Arguments

object
a HeatmapList-class object.

x
a Heatmap-class object or a HeatmapAnnotation-class object or a HeatmapList-class object.

direction
direction of the concatenation.

Details

There is a shortcut function +.AdditiveUnit.

Value

A HeatmapList-class object.

Author(s)

Zuguang Gu <z.gu@dkfz.de>

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]