adjust_heatmap_list-HeatmapList-method {ComplexHeatmap}	R Documentation

Adjust Heatmap List

Description

Adjust Heatmap List

Usage

## S4 method for signature 'HeatmapList'
adjust_heatmap_list(object)

Arguments

object
A HeatmapList-class object.

Details

This function adjusts settings in all other heatmaps according to the main heatmap. It also adjust the size of heatmap annotations to make them aligned nicely.

This function is only for internal use.

Author(s)

Zuguang Gu <z.gu@dkfz.de>

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]