anno_barplot {ComplexHeatmap}	R Documentation

Barplot Annotation

Description

Barplot Annotation

Usage

anno_barplot(x, baseline = 0, which = c("column", "row"), border = TRUE, bar_width = 0.6, beside = FALSE,
    gp = gpar(fill = "#CCCCCC"), ylim = NULL, extend = 0.05, axis = TRUE,
    axis_param = default_axis_param(which),
    add_numbers = FALSE, numbers_gp = gpar(fontsize = 8),
    numbers_rot = ifelse(which == "column", 45, 0), numbers_offset = unit(2, "mm"),
    width = NULL, height = NULL, ...)

Arguments

x
The value vector. The value can be a vector or a matrix. The length of the vector or the number of rows of the matrix is taken as the number of the observations of the annotation. If x is a vector, the barplots will be represented as stacked barplots.

baseline
baseline of bars. The value should be "min" or "max", or a numeric value. It is enforced to be zero for stacked barplots.

which
Whether it is a column annotation or a row annotation?

border
Wether draw borders of the annotation region?

bar_width
Relative width of the bars. The value should be smaller than one.

beside
When x is a matrix, will bars be positioned beside each other or as stacked bars?

gp
Graphic parameters for bars. The length of each graphic parameter can be 1, length of x if x is a vector, or number of columns of x is x is a matrix.

ylim
Data ranges. By default it is range(x) if x is a vector, or range(rowSums(x)) if x is a matrix.

extend
The extension to both side of ylim. The value is a percent value corresponding to ylim[2] - ylim[1].

axis
Whether to add axis?

axis_param
parameters for controlling axis. See default_axis_param for all possible settings and default parameters.

add_numbers
Whether to add numbers to the bars. It only works when x is a simple vector.

numbers_gp
Graphics parameters for the numbers.

numbers_rot
Rotation of numbers.

numbers_offset
Offset to the default positions (1mm away the top of the bars).

width
Width of the annotation. The value should be an absolute unit. Width is not allowed to be set for column annotation.

height
Height of the annotation. The value should be an absolute unit. Height is not allowed to be set for row annotation.

...
Other arguments.

Value

An annotation function which can be used in HeatmapAnnotation.

See Also

https://jokergoo.github.io/ComplexHeatmap-reference/book/heatmap-annotations.html#barplot_annotation

Examples

anno = anno_barplot(1:10)
draw(anno, test = "a vector")

m = matrix(runif(4*10), nc = 4)
m = t(apply(m, 1, function(x) x/sum(x)))
anno = anno_barplot(m, gp = gpar(fill = 2:5), bar_width = 1, height = unit(6, "cm"))
draw(anno, test = "proportion matrix")

[Package ComplexHeatmap version 2.10.0 Index]