anno_oncoprint_barplot {ComplexHeatmap}	R Documentation

Barplot Annotation for oncoPrint

Description

Barplot Annotation for oncoPrint

Usage

anno_oncoprint_barplot(type = NULL, which = c("column", "row"),
    bar_width = 0.6, beside = FALSE, ylim = NULL, show_fraction = FALSE, axis = TRUE,
    axis_param = if(which == "column") default_axis_param("column") else list(side = "top", labels_rot = 0),
    width = NULL, height = NULL, border = FALSE)

Arguments

type
A vector of the alteration types in the data. It can be a subset of all alteration types if you don't want to show them all.

which
Is it a row annotation or a column annotation?

bar_width
Width of the bars.

beside
Will bars be stacked or be positioned beside each other?

ylim
Data range.

show_fraction
Whether to show the numbers or the fractions?

axis
Whether draw axis?

axis_param
Parameters for controlling axis.

width
Width of the annotation.

height
Height of the annotation.

border
Whether draw the border?

Details

This annotation function should always be used with oncoPrint.

Author(s)

Zuguang Gu <z.gu@dkfz.de>

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]