bar3D {ComplexHeatmap}	R Documentation

Draw 3D bars

Description

Draw 3D bars

Usage

bar3D(x, y, w, h, l, theta = 60, default.units = "npc", fill = "white", col = "black")

Arguments

x
x coordinate of the center point in the bottom face.

y
y coordinate of the center point in the bottom face.

w
Width of the botton face.

h
Height of the botton face.

l
Length of the bars (in the z-direction).

theta
The angle for the projection.

default.units
Units.

fill
Filled colors for the bars.

col
Border colors.

Examples

grid.newpage()
bar3D(c(0.3, 0.7), 0.5, 0.2, 0.2, 0.2, fill = 2:3)

[Package ComplexHeatmap version 2.10.0 Index]