cluster_within_group {ComplexHeatmap}	R Documentation

Cluster within and between Groups

Description

Cluster within and between Groups

Usage

cluster_within_group(mat, factor)

Arguments

mat
A matrix where clustering is applied on columns.

factor
A categorical vector.

Details

The clustering is firstly applied in each group, then clustering is applied to group means. The within-group dendrograms and between-group dendrogram are finally connected by merge_dendrogram.

In the final dendrogram, the within group dendrograms are enforced to be flat lines to emphasize that the within group dendrograms have no sense to compare to between-group dendrogram.

Value

A dendrogram object. The order of columns can be retrieved by order.dendrogram.

Examples

m = matrix(rnorm(120), nc = 12)
colnames(m) = letters[1:12]
fa = rep(c("a", "b", "c"), times = c(2, 4, 6))
dend = cluster_within_group(m, fa)
grid.dendrogram(dend, test = TRUE)

[Package ComplexHeatmap version 2.10.0 Index]