comb_name {ComplexHeatmap}	R Documentation

Names of the Combination sets

Description

Names of the Combination sets

Usage

comb_name(m, readable = FALSE)

Arguments

m
A combination matrix returned by make_comb_mat.

readable
Whether the combination represents as e.g. "A&B&C".

Details

The name of the combination sets are formatted as a string of binary bits. E.g. for three sets of "a", "b", "c", the combination set with name "101" corresponds to select set a, not select set b and select set c. The definition of "select" depends on the value of mode from make_comb_mat.

Value

A vector of names of the combination sets.

Examples

set.seed(123)
lt = list(a = sample(letters, 10),
          b = sample(letters, 15),
          c = sample(letters, 20))
m = make_comb_mat(lt)
comb_name(m)
comb_name(m, readable = TRUE)

[Package ComplexHeatmap version 2.10.0 Index]