copy_all-AnnotationFunction-method {ComplexHeatmap}	R Documentation

Copy the AnnotationFunction Object

Description

Copy the AnnotationFunction Object

Usage

## S4 method for signature 'AnnotationFunction'
copy_all(object)

Arguments

object
The AnnotationFunction-class object.

Details

In AnnotationFunction-class, there is an environment which stores some external variables for the annotation function (specified by the var_import argument when constructing the AnnotationFunction-class object. This copy_all,AnnotationFunction-method hard copies all the variables into a new isolated environment.

The environment is at object@var_env.

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]