default_axis_param {ComplexHeatmap}	R Documentation

The Default Parameters for Annotation Axis

Description

The Default Parameters for Annotation Axis

Usage

default_axis_param(which)

Arguments

which
Whether it is for column annotation or row annotation?

Details

There are following parameters for the annotation axis:

at

The breaks of axis. By default it is automatically inferred.

labels

The corresponding axis labels.

labels_rot

The rotation of the axis labels.

gp

Graphc parameters of axis labels. The value should be a unit object.

side

If it is for column annotation, the value should only be one of left and right. If it is for row annotation, the value should only be one of top and bottom.

facing

Whether the axis faces to the outside of the annotation region or inside. Sometimes when appending more than one heatmaps, the axes of column annotations of one heatmap might overlap to the neighbouring heatmap, setting facing to inside may invoild it.

direction

The direction of the axis. Value should be "normal" or "reverse".

All the parameters are passed to annotation_axis_grob to construct an axis grob.

Examples

default_axis_param("column")
default_axis_param("row")

[Package ComplexHeatmap version 2.10.0 Index]