dendrogramGrob {ComplexHeatmap}	R Documentation

Grob for Dendrogram

Description

Grob for Dendrogram

Usage

dendrogramGrob(dend, facing = c("bottom", "top", "left", "right"),
    order = c("normal", "reverse"), gp = gpar())

Arguments

dend
A dendrogram object.

facing
Facing of the dendrogram.

order
If it is set to reverse, the first leaf is put on the right if the dendrogram is horizontal and it is put on the top if the dendrogram is vertical.

gp
Graphic parameters for the dendrogram segments. If any of col, lwd or lty is set in the edgePar attribute of a node, the corresponding value defined in gp will be overwritten for this node, so gp is like global graphic parameters for dendrogram segments.

Details

If dend has not been processed by adjust_dend_by_x, internally adjust_dend_by_x is called to add x attributes to each node/leaf.

Value

A grob object which is contructed by segmentsGrob.

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]