draw-AnnotationFunction-method {ComplexHeatmap}	R Documentation

Draw the AnnotationFunction Object

Description

Draw the AnnotationFunction Object

Usage

## S4 method for signature 'AnnotationFunction'
draw(object, index, k = 1, n = 1, test = FALSE, ...)

Arguments

object
The AnnotationFunction-class object.

index
Index of observations.

k
Current slice index.

n
Total number of slices.

test
Is it in test mode? The value can be logical or a text which is plotted as the title of plot.

...
Pass to viewport.

Details

Normally it is called internally by the SingleAnnotation-class.

When test is set to TRUE, the annotation graphic is directly drawn, which is generally for testing purpose.

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]