draw-Heatmap-method {ComplexHeatmap}	R Documentation

Draw a Single Heatmap

Description

Draw a Single Heatmap

Usage

## S4 method for signature 'Heatmap'
draw(object, internal = FALSE, test = FALSE, ...)

Arguments

object
A Heatmap-class object.

internal
If TRUE, it is only used inside the calling of draw,HeatmapList-method. It only draws the heatmap without legends where the legend will be drawn by draw,HeatmapList-method.

test
Only for testing. If it is TRUE, the heatmap body is directly drawn.

...
Pass to draw,HeatmapList-method.

Details

The function creates a HeatmapList-class object which only contains a single heatmap and call draw,HeatmapList-method to make the final heatmap.

There are some arguments which control the some settings of the heatmap such as legends. Please go to draw,HeatmapList-method for these arguments.

Value

A HeatmapList-class object.

Author(s)

Zuguang Gu <z.gu@dkfz.de>

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]