draw-HeatmapAnnotation-method {ComplexHeatmap}	R Documentation

Draw the Heatmap Annotations

Description

Draw the Heatmap Annotations

Usage

## S4 method for signature 'HeatmapAnnotation'
draw(object, index, k = 1, n = 1, ...,
    test = FALSE, anno_mark_param = list())

Arguments

object
A HeatmapAnnotation-class object.

index
A vector of indices.

k
The current slice index for the annotation if it is split.

n
Total number of slices.

...
Pass to viewport which contains all the annotations.

test
Is it in test mode? The value can be logical or a text which is plotted as the title of plot.

anno_mark_param
It contains specific parameters for drawing anno_mark and pass to the draw,SingleAnnotation-method.

Value

No value is returned.

Author(s)

Zuguang Gu <z.gu@dkfz.de>

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]