draw-SingleAnnotation-method {ComplexHeatmap}	R Documentation

Draw the Single Annotation

Description

Draw the Single Annotation

Usage

## S4 method for signature 'SingleAnnotation'
draw(object, index, k = 1, n = 1, test = FALSE,
    anno_mark_param = list())

Arguments

object
A SingleAnnotation-class object.

index
A vector of indices.

k
The index of the slice.

n
Total number of slices. k and n are used to adjust annotation names. E.g. if k is 2 and n is 3, the annotation names are not drawn.

test
Is it in test mode? The value can be logical or a text which is plotted as the title of plot.

anno_mark_param
It contains specific parameters for drawing anno_mark.

Value

No value is returned.

Author(s)

Zuguang Gu <z.gu@dkfz.de>

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]