draw_annotation_legend-HeatmapList-method {ComplexHeatmap}	R Documentation

Draw legends for All Annotations

Description

Draw legends for All Annotations

Usage

## S4 method for signature 'HeatmapList'
draw_annotation_legend(object, legend_list = list(), ...)

Arguments

object
A HeatmapList-class object.

legend_list
A list of self-defined legends, should be wrapped into grob objects. It is normally constructed by Legend.

...
Other arguments.

Details

We call the "annotation legends" as the secondary legends. For horizontal heamtap list, the legends are those from all top/bottom annotations, and for vertical heatmap list, the legends are those from all left/right annotations.

A viewport is created which contains annotation legends.

This function is only for internal use.

Value

This function returns no value.

Author(s)

Zuguang Gu <z.gu@dkfz.de>

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]