draw_title-Heatmap-method {ComplexHeatmap}	R Documentation

Draw Heatmap Title

Description

Draw Heatmap Title

Usage

## S4 method for signature 'Heatmap'
draw_title(object,
    which = c("row", "column"), k = 1, ...)

Arguments

object
A Heatmap-class object.

which
Is title put on the row or on the column of the heatmap?

k
Slice index.

...
Pass to viewport which includes the complete heatmap title.

Details

A viewport is created which contains heatmap title.

This function is only for internal use.

Value

This function returns no value.

Author(s)

Zuguang Gu <z.gu@dkfz.de>

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]