draw_title-HeatmapList-method {ComplexHeatmap}	R Documentation

Draw Heatmap List Title

Description

Draw Heatmap List Title

Usage

## S4 method for signature 'HeatmapList'
draw_title(object,
    which = c("column", "row"))

Arguments

object
A HeatmapList-class object.

which
Is it a row title or a column title.

Details

A viewport is created which contains heatmap list title.

This function is only for internal use.

Value

This function returns no value.

Author(s)

Zuguang Gu <z.gu@dkfz.de>

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]