grid.annotation_axis {ComplexHeatmap}	R Documentation

Draw Annotation Axis

Description

Draw Annotation Axis

Usage

grid.annotation_axis(at = NULL, labels = at, labels_rot = 0, gp = gpar(),
    side = "left", facing = "outside", direction = "normal")

Arguments

at
Break values. If it is not specified, it is inferred from data scale in current viewport.

labels
Corresponding labels.

labels_rot
Rotations of labels.

gp
Graphic parameters.

side
side of the axis of the annotation viewport.

facing
Facing of the axis.

direction
direction of the axis. Value should be "normal" or "reverse".

Details

It uses annotation_axis_grob to construct the grob object, then use grid.draw to draw the axis.

Examples

# See examples in `annotation_axis_grob`
NULL

[Package ComplexHeatmap version 2.10.0 Index]