is_abs_unit {ComplexHeatmap}	R Documentation

Test Whether it is an Absolute Unit

Description

Test Whether it is an Absolute Unit

Usage

is_abs_unit(u)

Arguments

u
A unit object.

Details

Besides the normal absolute units (e.g. "mm", "inches"), this function simply assumes grob objects as absolute units.

For a complex unit which is combination of different units, it is absolute only if all units included are absolute units.

Value

A logical value.

Author(s)

Zuguang Gu <z.gu@dkfz.de>

Examples

is_abs_unit(unit(1, "mm"))
is_abs_unit(unit(1, "npc"))
is_abs_unit(grobWidth(textGrob("foo")))
is_abs_unit(unit(1, "mm") + unit(1, "npc"))

[Package ComplexHeatmap version 2.10.0 Index]