make_layout-Heatmap-method {ComplexHeatmap}	R Documentation

Make the Layout of a Single Heatmap

Description

Make the Layout of a Single Heatmap

Usage

## S4 method for signature 'Heatmap'
make_layout(object)

Arguments

object
A Heatmap-class object.

Details

The layout of the single heatmap will be established by setting the size of each heatmap component. Also how to make graphics for heatmap components will be recorded by saving as functions.

Whether to apply row clustering or column clustering affects the layout, so clustering should be applied first by prepare,Heatmap-method before making the layout.

This function is only for internal use.

Value

A Heatmap-class object.

Author(s)

Zuguang Gu <z.gu@dkfz.de>

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]