Make Layout for the Heatmap List

Make Layout for the Heatmap List

# S4 method for HeatmapList
make_layout(object,

    row_title = character(0),
    row_title_side = c("left", "right"),
    row_title_gp = gpar(fontsize = 14),
    column_title = character(0),
    column_title_side = c("top", "bottom"),
    column_title_gp = gpar(fontsize = 14),

    heatmap_legend_side = c("right", "left", "bottom", "top"),
    merge_legends = FALSE,
    show_heatmap_legend = TRUE,
    heatmap_legend_list = list(),
    annotation_legend_side = c("right", "left", "bottom", "top"),
    show_annotation_legend = TRUE,
    annotation_legend_list = list(),

    ht_gap = unit(2, "mm"),

    main_heatmap = which(sapply(object@ht_list, inherits, "Heatmap"))[1],
    padding = GLOBAL_PADDING,

    auto_adjust = TRUE,
    row_dend_side = c("original", "left", "right"),
    row_sub_title_side = c("original", "left", "right"),
    column_dend_side = c("original", "top", "bottom"),
    column_sub_title_side = c("original", "top", "bottom"),

    row_gap = NULL,
    cluster_rows = NULL,
    cluster_row_slices = NULL,
    clustering_distance_rows = NULL,
    clustering_method_rows = NULL,
    row_dend_width = NULL,
    show_row_dend = NULL,
    row_dend_reorder = NULL,
    row_dend_gp = NULL,
    row_order = NULL,
    row_km = NULL,
    row_km_repeats = NULL,
    row_split = NULL,
    height = NULL,
    heatmap_height = NULL,

    column_gap = NULL,
    cluster_columns = NULL,
    cluster_column_slices = NULL,
    clustering_distance_columns = NULL,
    clustering_method_columns = NULL,
    column_dend_width = NULL,
    show_column_dend = NULL,
    column_dend_reorder = NULL,
    column_dend_gp = NULL,
    column_order = NULL,
    column_km = NULL,
    column_km_repeats = NULL,
    column_split = NULL,
    width = NULL,
    heatmap_width = NULL)

Arguments

object
A HeatmapList-class object.

row_title
Title on the row.

row_title_side
Will the title be put on the left or right of the heatmap list?

row_title_gp
Graphic parameters for the row title.

column_title
Title on the column.

column_title_side
Will the title be put on the top or bottom of the heatmap?

column_title_gp
Graphic parameters for the column title.

heatmap_legend_side
Side of the heatmap legends.

merge_legends
Whether to put heatmap legends and annotation legends together. By default they are put in different viewports.

show_heatmap_legend
Whether show heatmap legends.

heatmap_legend_list
A list of self-defined legends, should be wrapped into a list of grob objects. Normally they are constructed by Legend.

annotation_legend_side
Side of annotation legends.

show_annotation_legend
Whether show annotation legends.

annotation_legend_list
A list of self-defined legends, should be wrapped into a list of grob objects. Normally they are constructed by Legend.

ht_gap
Gap between heatmaps, should be a unit object. It can be a vector of length 1 or the number of heamtaps/annotations.

main_heatmap
Name or index for the main heatmap.

padding
Padding of the whole plot. The four values correspond to the bottom, left, top and right paddings.

auto_adjust
whether apply automatic adjustment? The auto-adjustment includes turning off dendrograms, titles and row/columns for non-main heatmaps.

row_dend_side
If auto-adjustment is on, to put the row dendrograms of the main heatmap to the most left side of the heatmap list or the most right side?

row_sub_title_side
There can be sub titles generated by the splitting of heatmaps. Similar setting as row_dend_side.

column_dend_side
Similar setting as row_dend_side.

column_sub_title_side
Similar setting as row_sub_title_side.

row_gap
Overwrite the corresponding setting in the main heatmap.

cluster_rows
Overwrite the corresponding setting in the main heatmap.

cluster_row_slices
Overwrite the corresponding setting in the main heatmap.

clustering_distance_rows
Overwrite the corresponding setting in the main heatmap.

clustering_method_rows
Overwrite the corresponding setting in the main heatmap.same setting as in Heatmap, if it is specified, clustering_method_rows in main heatmap is ignored.

row_dend_width
Overwrite the corresponding setting in the main heatmap.

show_row_dend
same Overwrite the corresponding setting in the main heatmap.

row_dend_reorder
Overwrite the corresponding setting in the main heatmap.

row_dend_gp
Overwrite the corresponding setting in the main heatmap.

row_order
Overwrite the corresponding setting in the main heatmap.

row_km
Overwrite the corresponding setting in the main heatmap.

row_km_repeats
Overwrite the corresponding setting in the main heatmap.

row_split
Overwrite the corresponding setting in the main heatmap.

height
Overwrite the corresponding setting in the main heatmap.

heatmap_height
Overwrite the corresponding setting in the main heatmap.

column_gap
Overwrite the corresponding setting in the main heatmap.

cluster_columns
Overwrite the corresponding setting in the main heatmap.

cluster_column_slices
Overwrite the corresponding setting in the main heatmap.

clustering_distance_columns
Overwrite the corresponding setting in the main heatmap.

clustering_method_columns
Overwrite the corresponding setting in the main heatmap.

column_dend_width
column Overwrite the corresponding setting in the main heatmap.

show_column_dend
Overwrite the corresponding setting in the main heatmap.

column_dend_reorder
Overwrite the corresponding setting in the main heatmap.

column_dend_gp
Overwrite the corresponding setting in the main heatmap.

column_order
Overwrite the corresponding setting in the main heatmap.

column_km
Overwrite the corresponding setting in the main heatmap.

column_km_repeats
Overwrite the corresponding setting in the main heatmap.

column_split
Overwrite the corresponding setting in the main heatmap.

width
Overwrite the corresponding setting in the main heatmap.

heatmap_width
Overwrite the corresponding setting in the main heatmap.

Details

It sets the size of each component of the heatmap list and adjusts graphic parameters for each heatmap if necessary.

This function is only for internal use.

Value

A HeatmapList-class object in which settings for all heatmap are adjusted.

Examples

# There is no example
NULL

#> NULL