max_text_height {ComplexHeatmap}	R Documentation

Maximum Height of Text

Description

Maximum Height of Text

Usage

max_text_height(text, gp = gpar(), rot = 0)

Arguments

text
A vector of text.

gp
Graphic parameters for text.

rot
Rotation of the text, scalar.

Details

It simply calculates maximum height of a list of textGrob objects.

Note it ignores the text rotation.

Value

A unit object.

Author(s)

Zuguang Gu <z.gu@dkfz.de>

See Also

max_text_width calculates the maximum width of a text vector.

Examples

x = c("a", "b\nb", "c\nc\nc")
max_text_height(x, gp = gpar(fontsize = 10))

[Package ComplexHeatmap version 2.10.0 Index]