max_text_width {ComplexHeatmap}	R Documentation

Maximum Width of Text

Description

Maximum Width of Text

Usage

max_text_width(text, gp = gpar(), rot = 0)

Arguments

text
A vector of text.

gp
Graphic parameters for text.

rot
Rotation of the text, scalar.

Details

It simply calculates maximum width of a list of textGrob objects.

Note it ignores the text rotation.

Value

A unit object which is in "mm".

Author(s)

Zuguang Gu <z.gu@dkfz.de>

See Also

max_text_height calculates the maximum height of a text vector.

Examples

x = c("a", "bb", "ccc")
max_text_width(x, gp = gpar(fontsize = 10))

[Package ComplexHeatmap version 2.10.0 Index]