normalize_comb_mat {ComplexHeatmap}	R Documentation

Normalize a list of combination matrice

Description

Normalize a list of combination matrice

Usage

normalize_comb_mat(..., full_comb_sets = FALSE, complement_set = FALSE)

Arguments

...
Combination matrices.

full_comb_sets
Whether the combination matrices contain the full sets of combination sets?

complement_set
Whether the combination matrices also contain the complement set?

Details

It normalizes a list of combination matrice to make them have same number and order of sets and combination sets.

The sets (by set_name) from all combination matrice should be the same.

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]