Get Values in a Matrix by Pair-wise Indices

Get Values in a Matrix by Pair-wise Indices

pindex(m, i, j)

Arguments

m
A matrix or a 3-dimension array.

i
Row indices or the indices in the first dimension.

j
Column indicies or the indices in the second dimension.

Value

If m is a matrix, the value returned is a vector c(m[i1, j1], m[i2, j2], ...)`.

If m is an array, the value returned is a matrix rbind(m[i1, j1, ], m[i2, j2, ], ...)`.

Examples

m = matrix(rnorm(100), 10)
m2 = m[m > 0]
ind = do.call("rbind", lapply(1:10, function(ci) {
    i = which(m[, ci] > 0)
    cbind(i = i, j = rep(ci, length(i)))
}))
pindex(m, ind[, 1], ind[, 2])

#>  [1] 0.73649596 0.38602657 0.11814451 0.13403865 0.22101947 1.64084617
#>  [7] 0.16806538 1.16838387 1.05418102 1.14526311 2.00248273 0.06670087
#> [13] 1.86685184 0.02098359 1.24991457 0.33117917 0.21198043 1.23667505
#> [19] 2.03757402 1.30117599 0.75677476 0.70352390 1.68443571 0.91139129
#> [25] 0.23743027 1.21810861 0.66082030 0.68374552 0.63296071 1.33551762
#> [31] 0.00729009 1.01755864 1.51921771 0.37738797 0.86577940 0.62418747
#> [37] 0.95900538 1.67105483 0.05601673 0.09932759 1.01494317 0.11663728
#> [43] 0.33390294 0.41142992 2.57145815 0.65119328 0.27376649 1.02467323
#> [49] 0.81765945 0.37816777 0.85692301 2.41677335 0.82537986 0.51013255
#> [55] 0.14447570

identical(pindex(m, ind[, 1], ind[, 2]), m[m > 0])

#> [1] TRUE

# 3d array
arr = array(1:27, dim = c(3, 3, 3))
pindex(arr, 1:2, 2:3)

#>      [,1] [,2] [,3]
#> [1,]    4   13   22
#> [2,]    8   17   26

identical(pindex(arr, 1:2, 2:3),
   rbind(arr[1, 2, ], arr[2, 3, ]))

#> [1] TRUE