packLegend {ComplexHeatmap}	R Documentation

Pack Legends

Description

Pack Legends

Usage

packLegend(..., gap = unit(4, "mm"), row_gap = unit(4, "mm"), column_gap = unit(4, "mm"),
    direction = c("vertical", "horizontal"),
    max_width = NULL, max_height = NULL, list = NULL)

Arguments

...
A list of objects returned by Legend.

gap
Gap between two neighbouring legends. The value is a unit object with length of one. It is the same as row_gap if the direction if vertial and the same as column_gap if the direction is horizontal.

row_gap
Horizontal gaps between legends.

column_gap
Vertical gaps between legends.

direction
The direction to arrange legends.

max_width
The maximal width of the total packed legends. It only works for horizontal arrangement. If the total width of the legends exceeds it, the legends will be arranged into multiple rows.

max_height
Similar as max_width, but for the vertical arrangment of legends.

list
The list of legends can be specified as a list.

Value

A Legends-class object.

See Also

https://jokergoo.github.io/ComplexHeatmap-reference/book/legends.html#a-list-of-legends

Examples

require(circlize)
col_fun = colorRamp2(c(0, 0.5, 1), c("blue", "white", "red"))
lgd1 = Legend(at = 1:6, legend_gp = gpar(fill = 1:6), title = "legend1")
lgd2 = Legend(col_fun = col_fun, title = "legend2", at = c(0, 0.25, 0.5, 0.75, 1))
pd = packLegend(lgd1, lgd2)
draw(pd, test = "two legends")
pd = packLegend(lgd1, lgd2, direction = "horizontal")
draw(pd, test = "two legends packed horizontally")

[Package ComplexHeatmap version 2.10.0 Index]