%v% {ComplexHeatmap}	R Documentation

Vertically Add Heatmaps or Annotations to a Heatmap List

Description

Vertically Add Heatmaps or Annotations to a Heatmap List

Usage

x %v% y

Arguments

x
A Heatmap-class object, a HeatmapAnnotation-class object or a HeatmapList-class object.

y
A Heatmap-class object, a HeatmapAnnotation-class object or a HeatmapList-class object.

Details

It is only a helper function. It actually calls add_heatmap,Heatmap-method, add_heatmap,HeatmapList-method or add_heatmap,HeatmapAnnotation-method depending on the class of the input objects.

The HeatmapAnnotation-class object to be added should only be column annotations.

x and y can also be NULL.

Value

A HeatmapList-class object.

Author(s)

Zuguang Gu <z.gu@dkfz.de>

See Also

+.AdditiveUnit operator is used for horizontal heatmap list.

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]