re_size-HeatmapAnnotation-method {ComplexHeatmap}	R Documentation

Resize the Width or Height of Heatmap Annotations

Description

Resize the Width or Height of Heatmap Annotations

Usage

## S4 method for signature 'HeatmapAnnotation'
re_size(object,
    annotation_height = NULL,
    annotation_width = NULL,
    height = NULL,
    width = NULL,
    simple_anno_size = object@param$simple_anno_size,
    simple_anno_size_adjust = object@param$simple_anno_size_adjust)

Arguments

object
A HeatmapAnnotation-class object.

annotation_height
A vector of of annotation heights in unit class.

annotation_width
A vector of of annotation widths in unit class.

height
The height of the complete heatmap annotation.

width
The width of the complete heatmap annotation.

simple_anno_size
The size of one line of the simple annotation.

simple_anno_size_adjust
Whether adjust the size of the simple annotation?

Details

The function only adjust height for column annotations and width for row annotations.

The basic rules are (take height and annotation_height for example:

1. If annotation_height is set and all annotation_height are absolute units, height is ignored. 2. If annotation_height contains non-absolute units, height also need to be set and the non-absolute units should be set in a simple form such as 1:10 or unit(1, "null"). 3. simple_anno_size is only used when annotation_height is NULL. 4. If only height is set, non-simple annotation is adjusted while keeps simple anntation unchanged. 5. If only height is set and all annotations are simple annotations, all anntations are adjusted, and simple_anno_size is disabled. 6. If simple_anno_size_adjust is FALSE, the size of the simple annotations will not change.

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]