size.AnnotationFunction {ComplexHeatmap}	R Documentation

Size of the AnnotationFunction Object

Description

Size of the AnnotationFunction Object

Usage

## S3 method for class 'AnnotationFunction'
size(x, ...)

Arguments

x
The AnnotationFunction-class object.

...
Other arguments.

Details

It returns the width if it is a row annotation and the height if it is a column annotation.

Internally used.

Examples

anno = anno_points(1:10)
ComplexHeatmap:::size(anno)
anno = anno_points(1:10, which = "row")
ComplexHeatmap:::size(anno)

[Package ComplexHeatmap version 2.10.0 Index]