size.SingleAnnotation {ComplexHeatmap}	R Documentation

Size of the SingleAnnotation Object

Description

Size of the SingleAnnotation Object

Usage

## S3 method for class 'SingleAnnotation'
size(x, ...)

Arguments

x
The SingleAnnotation-class object.

...
Other arguments.

Details

It returns the width if it is a row annotation and the height if it is a column annotation.

Internally used.

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]