unify_mat_list {ComplexHeatmap}	R Documentation

Unify a List of Matrix

Description

Unify a List of Matrix

Usage

unify_mat_list(mat_list, default = 0)

Arguments

mat_list
A list of matrix. All of them should have dimension names.

default
Default values for the newly added rows and columns.

Details

All matrix will be unified to have same row names and column names.

Value

A list of matrix

Author(s)

Zuguang Gu <z.gu@dkfz.de>

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]