upset_left_annotation {ComplexHeatmap}	R Documentation

UpSet Left Annotation

Description

UpSet Left Annotation

Usage

upset_left_annotation(m,
    gp = gpar(fill = "black"),
    axis_param = list(direction = "reverse"),
    width = unit(ifelse(set_on_rows, 2, 3), "cm"),
    show_annotation_name = TRUE,
    annotation_name_gp = gpar(),
    annotation_name_offset = NULL,
    annotation_name_side = "bottom",
    annotation_name_rot = NULL,
    ...)

Arguments

m
A combination matrix which is as same as the one for UpSet.

gp
Graphic parameters for bars.

axis_param
Parameters for axis.

width
Width of the left annotation.

show_annotation_name
Whether show annotation names?

annotation_name_gp
Graphic parameters for anntation names.

annotation_name_offset
Offset to the annotation name, a unit object.

annotation_name_side
Side of the annotation name.

annotation_name_rot
Rotation of the annotation name, it can only take values in c(00, 90, 180, 270).

...
Passed to anno_barplot, e.g. to set add_numbers.

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]