upset_right_annotation {ComplexHeatmap}	R Documentation

Default UpSet Right Annotation

Description

Default UpSet Right Annotation

Usage

upset_right_annotation(m,
    gp = gpar(fill = "black"),
    width = unit(ifelse(set_on_rows, 2, 3), "cm"),
    show_annotation_name = TRUE,
    annotation_name_gp = gpar(),
    annotation_name_offset = NULL,
    annotation_name_side = "bottom",
    annotation_name_rot = NULL,
    ...)

Arguments

m
A combination matrix which is as same as the one for UpSet.

gp
Graphic parameters for bars.

width
Width of the right annotation.

show_annotation_name
Whether show annotation names?

annotation_name_gp
Graphic parameters for anntation names.

annotation_name_offset
Offset to the annotation name, a unit object.

annotation_name_side
Side of the annotation name.

annotation_name_rot
Rotation of the annotation name, it can only take values in c(00, 90, 180, 270).

...
Passed to anno_barplot, e.g. to set add_numbers.

Details

The default right annotation is actually barplot implemented by anno_barplot. For how to set the right annotation or left annotation in UpSet, please refer to UpSet.

If you want to use decorate_annotation function, the annotation name for the "sets" is set_size and the annotation name for the "intersection sets" are intersection_size and if under the union mode, it is union_size.

Examples

# There is no example
NULL

[Package ComplexHeatmap version 2.10.0 Index]