adjust_dend_by_x {ComplexHeatmap}	R Documentation

Adjust the Positions of nodes/leaves in the Dendrogram

Description

Adjust the Positions of nodes/leaves in the Dendrogram

Usage

adjust_dend_by_x(dend, leaf_pos = 1:nobs(dend)-0.5)

Arguments

dend
A dendrogram object.

leaf_pos
A vector of positions of leaves. The value can also be a unit object.

Details

The positions of nodes stored as x attribute are recalculated based on the new positions of leaves.

By default, the position of leaves are at 0.5, 1.5, ..., n-0.5.

Examples

m = matrix(rnorm(100), 10)
dend = as.dendrogram(hclust(dist(m)))
dend = adjust_dend_by_x(dend, sort(runif(10)))
str(dend)
dend = adjust_dend_by_x(dend, unit(1:10, "cm"))
str(dend)

[Package ComplexHeatmap version 2.10.0 Index]