ht_opt {ComplexHeatmap}	R Documentation

Global Options for Heatmaps

Description

Global Options for Heatmaps

Usage

ht_opt(..., RESET = FALSE, READ.ONLY = NULL, LOCAL = FALSE, ADD = FALSE)

Arguments

...
Options, see 'Details' section.

RESET
Reset all the option values.

READ.ONLY
Please ignore this argument.

LOCAL
Please ignore this argument.

ADD
Please ignore this argument.

Details

You can set some parameters for all heatmaps/annotations simultaneously by this global function. Pleast note you should put it before your heatmap code and reset all option values after drawing the heatmaps to get rid of affecting next heatmap.

There are following parameters to control all heatmaps:

heatmap_row_names_gp

set row_names_gp in all Heatmap.

heatmap_column_names_gp

set column_names_gp in all Heatmap.

heatmap_row_title_gp

set row_title_gp in all Heatmap.

heatmap_column_title_gp

set column_title_gp in all Heatmap.

heatmap_border

set border in all Heatmap.

Following parameters control the legends:

legend_title_gp

set title_gp in all heatmap legends and annotation legends.

legend_title_position

set title_position in all heatmap legends and annotation legends.

legend_labels_gp

set labels_gp in all heatmap legends and annotation legends.

legend_grid_width

set grid_width in all heatmap legends and annotation legends.

legend_grid_height

set grid_height in all heatmap legends and annotation legends.

legend_border

set border in all heatmap legends and annotation legends.

legend_gap

Gap between legends. The value should be a vector of two units. One for gaps between vertical legends and one for the horizontal legends. If only one single unit is specified, the same gap set for the vertical and horizontal legends.

Following parameters control heatmap annotations:

annotation_border

border in all HeatmapAnnotation.

simple_anno_size

size for the simple annotation.

Following parameters control the space between heatmap components:

DENDROGRAM_PADDING

space bewteen dendrograms and heatmap body.

DIMNAME_PADDING

space between row/column names and heatmap body.

TITLE_PADDING

space between row/column titles and heatmap body. The value can have length of two which corresponds to the botton and top padding.

COLUMN_ANNO_PADDING

space between column annotations and heatmap body.

ROW_ANNO_PADDING

space between row annotations and heatmap body.

HEATMAP_LEGEND_PADDING

space between heatmap legends and heatmaps

ANNOTATION_LEGEND_PADDING

space between annotation legends and heatmaps

Other parameters:

fast_hclust

whether use hclust to speed up clustering?

show_parent_dend_line

when heatmap is split, whether to add a dashed line to mark parent dendrogram and children dendrograms?

You can get or set option values by the traditional way (like options) or by $ operator:

    # to get option values
    ht_opt("heatmap_row_names_gp")
    ht_opt$heatmap_row_names_gp

    # to set option values
    ht_opt("heatmap_row_names_gp" = gpar(fontsize = 8))
    ht_opt$heatmap_row_names_gp = gpar(fontsize = 8)  

Reset to the default values by ht_opt(RESET = TRUE).

Author(s)

Zuguang Gu <z.gu@dkfz.de>

Examples

ht_opt

[Package ComplexHeatmap version 2.10.0 Index]