list_to_matrix {ComplexHeatmap}	R Documentation

Convert a List of Sets to a Binary Matrix

Description

Convert a List of Sets to a Binary Matrix

Usage

list_to_matrix(lt, universal_set = NULL)

Arguments

lt
A list of vectors.

universal_set
The universal set.

Details

It converts the list which have m sets to a binary matrix with n rows and m columns where n is the size of universal set.

Examples

set.seed(123)
lt = list(a = sample(letters, 5),
          b = sample(letters, 10),
          c = sample(letters, 15))
list_to_matrix(lt)
list_to_matrix(lt, universal_set = letters)

[Package ComplexHeatmap version 2.10.0 Index]